add sub directories feature
By default no sub directories are created in any users home. Pass a list
of sub directories to $sftp_jail::sub_dirs. This list is used as a
default list of sub directories for all users.

If you want to create an individual list of sub directories in one users
home, pass a list of sub directories to the sub_dirs parameter, when
creating the user with sftp_jail::jail or sftp_jail::user respectively.

Optionally it is possible to enable merging. If enabled, the users sub
directory list is merged with the default sub directory list. This is
usefull, when multiple or all users do have some sub directories in
common. Merging can be enabled globally by setting
$sftp_jail::merge_subdirs to true. Similar to the list of sub
directories, it is possible to enable or disable merging on a per user
basis.